change floats to doubles
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@2835 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/trafficmanager.hpp
@@ -162,8 +162,8 @@ protected:
   int   _warmup_time;
   int   _drain_time;
 
-  float _load;
-  float _flit_rate;
+  double _load;
+  double _flit_rate;
 
   int   _packet_size;
